Move FeedPublisher class to more suitable directory path
