+++ src/main/java/org/opentripplanner/apis/gtfs/datafetchers/FeedImpl.java
@@ -9,11 +9,11 @@ import java.util.stream.Collectors;
 import org.opentripplanner.apis.gtfs.GraphQLRequestContext;
 import org.opentripplanner.apis.gtfs.generated.GraphQLDataFetchers;
 import org.opentripplanner.apis.gtfs.generated.GraphQLTypes;
+import org.opentripplanner.apis.gtfs.model.FeedPublisher;
 import org.opentripplanner.routing.alertpatch.EntitySelector;
 import org.opentripplanner.routing.alertpatch.TransitAlert;
 import org.opentripplanner.routing.services.TransitAlertService;
 import org.opentripplanner.transit.model.organization.Agency;
-import org.opentripplanner.transit.model.organization.FeedPublisher;
 import org.opentripplanner.transit.service.TransitService;
 
 public class FeedImpl implements GraphQLDataFetchers.GraphQLFeed {
